hrmsBackend/requirements.txt
asgiref==3.11.0
asttokens==3.0.1
decorator==5.2.1
dj-database-url==3.1.0
Django==4.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==2.2.1
gunicorn==21.2.0
ipdb==0.13.13
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
packaging==26.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.5.3
whitenoise==6.11.0
